cleaned up whitelist & user
