+++ Railgun/Commands/Inactivity/InactivityWhitelist.cs
@@ -1,0 +1,11 @@
+﻿using Finite.Commands;
+using Railgun.Core;
+
+namespace Railgun.Commands.Inactivity
+{
+    public partial class Inactivity
+    {
+        [Alias("whitelist")]
+        public partial class Whitelist : SystemBase { }
+    }
+}
